requirements.txt
groq
python-dotenv
SpeechRecognition
pyttsx3
pyaudio
